Created TouchListener class which implements RecyclerView.OnItemTouchListener.
